Extract game parsing validation to another class
