+++ game-core/src/main/java/games/strategy/engine/data/gameparser/GameParser.java
@@ -39,7 +39,6 @@ import games.strategy.triplea.attachments.TechAbilityAttachment;
 import games.strategy.triplea.attachments.UnitAttachment;
 import games.strategy.triplea.delegate.GenericTechAdvance;
 import games.strategy.triplea.delegate.TechAdvance;
-import games.strategy.triplea.formatter.MyFormatter;
 import java.io.InputStream;
 import java.net.URI;
 import java.util.ArrayList;
@@ -187,10 +186,11 @@ public final class GameParser {
     // set default tech attachments (comes after we parse all technologies, parse all attachments,
     // and parse all game
     // options/properties)
-    checkThatAllUnitsHaveAttachments(data);
     TechAbilityAttachment.setDefaultTechnologyAttachments(data);
     try {
-      validate();
+      final var validation = new GameParsingValidation(data);
+      validation.validate();
+      validation.checkThatAllUnitsHaveAttachments();
     } catch (final Exception e) {
       throw new GameParseException(
           String.format("map name: '%s', validation failed: %s", mapName, e.getMessage()), e);
@@ -221,54 +221,6 @@ public final class GameParser {
               ClientContext.engineVersion(),
               mapMinimumEngineVersion.toString(),
               data.getGameName()));
-    }
-  }
-
-  private void validate() throws GameParseException {
-    // validate unit attachments
-    for (final UnitType u : data.getUnitTypeList()) {
-      validateAttachments(u);
-    }
-    for (final Territory t : data.getMap()) {
-      validateAttachments(t);
-    }
-    for (final Resource r : data.getResourceList().getResources()) {
-      validateAttachments(r);
-    }
-    for (final GamePlayer r : data.getPlayerList().getPlayers()) {
-      validateAttachments(r);
-    }
-    for (final RelationshipType r : data.getRelationshipTypeList().getAllRelationshipTypes()) {
-      validateAttachments(r);
-    }
-    for (final TerritoryEffect r : data.getTerritoryEffectList().values()) {
-      validateAttachments(r);
-    }
-    for (final TechAdvance r : data.getTechnologyFrontier().getTechs()) {
-      validateAttachments(r);
-    }
-    // if relationships are used, every player should have a relationship with every other player
-    validateRelationships();
-  }
-
-  private void validateRelationships() throws GameParseException {
-    // for every player
-    for (final GamePlayer player : data.getPlayerList()) {
-      // in relation to every player
-      for (final GamePlayer player2 : data.getPlayerList()) {
-        // See if there is a relationship between them
-        if ((data.getRelationshipTracker().getRelationshipType(player, player2) == null)) {
-          // or else throw an exception!
-          throw new GameParseException(
-              "No relation set for: " + player.getName() + " and " + player2.getName());
-        }
-      }
-    }
-  }
-
-  private void validateAttachments(final Attachable attachable) throws GameParseException {
-    for (final IAttachment a : attachable.getAttachments().values()) {
-      a.validate(data);
     }
   }
 
@@ -932,8 +884,8 @@ public final class GameParser {
             nestedForeach.size() == 2
                 ? Splitter.on(":").splitToList(nestedForeach.get(1))
                 : List.of();
-        validateForeachVariables(foreachVariables1, variables, foreach);
-        validateForeachVariables(foreachVariables2, variables, foreach);
+        GameParsingValidation.validateForeachVariables(foreachVariables1, variables, foreach);
+        GameParsingValidation.validateForeachVariables(foreachVariables2, variables, foreach);
         final int length1 = variables.get(foreachVariables1.get(0)).size();
         for (int i = 0; i < length1; i++) {
           final Map<String, String> foreachMap1 =
@@ -950,27 +902,6 @@ public final class GameParser {
             }
           }
         }
-      }
-    }
-  }
-
-  private void validateForeachVariables(
-      final List<String> foreachVariables,
-      final Map<String, List<String>> variables,
-      final String foreach)
-      throws GameParseException {
-    if (foreachVariables.isEmpty()) {
-      return;
-    }
-    if (!variables.keySet().containsAll(foreachVariables)) {
-      throw new GameParseException("Attachment has invalid variables in foreach: " + foreach);
-    }
-    final int length = variables.get(foreachVariables.get(0)).size();
-    for (final String foreachVariable : foreachVariables) {
-      final List<String> foreachValue = variables.get(foreachVariable);
-      if (length != foreachValue.size()) {
-        throw new GameParseException(
-            "Attachment foreach variables must have same number of elements: " + foreach);
       }
     }
   }
@@ -1196,22 +1127,6 @@ public final class GameParser {
       final Resource resource = getResource(current.getAttribute("resource"));
       final int quantity = Integer.parseInt(current.getAttribute("quantity"));
       player.getResources().addResource(resource, quantity);
-    }
-  }
-
-  private void checkThatAllUnitsHaveAttachments(final GameData data) throws GameParseException {
-    final Collection<UnitType> errors = new ArrayList<>();
-    for (final UnitType ut : data.getUnitTypeList().getAllUnitTypes()) {
-      final UnitAttachment ua = UnitAttachment.get(ut);
-      if (ua == null) {
-        errors.add(ut);
-      }
-    }
-    if (!errors.isEmpty()) {
-      throw new GameParseException(
-          data.getGameName()
-              + " does not have unit attachments for: "
-              + MyFormatter.defaultNamedToTextList(errors));
     }
   }
 }
